+++ adapter/pkg/mgw/mgw.go
@@ -33,7 +33,7 @@ import (
 	"github.com/fsnotify/fsnotify"
 	"github.com/wso2/micro-gw/config"
 	logger "github.com/wso2/micro-gw/loggers"
-	"github.com/wso2/micro-gw/pkg/messagelisteners"
+	"github.com/wso2/micro-gw/pkg/messaging"
 	xds "github.com/wso2/micro-gw/pkg/xds"
 	"google.golang.org/grpc"
 )
@@ -116,7 +116,7 @@ func Run(conf *config.Config) {
 	var enableEventHub bool
 	enableEventHub = conf.ControlPlane.EventHub.Enable
 	if enableEventHub {
-		go messagelisteners.ProcessEvents(conf)
+		go messaging.ProcessEvents(conf)
 	}
 OUTER:
 	for {
